LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/at-container-style-serialization-expected.txt


FAIL Normalize spaces assert_equals: expected "style(--foo: bar)" but got "style( --foo:bar)"
FAIL Empty declaration value - spaces assert_equals: expected "style(--foo: )" but got "STyle(--foo: )"
FAIL Empty declaration value assert_equals: expected "style(--foo: )" but got "STyle(--foo:)"
PASS Missing declaration value
FAIL Unknown CSS property after 'or' assert_equals: expected "style((--FOO: BAR) or ( prop: val ))" but got "style( ( --FOO: BAR) OR ( prop: val ) )"
PASS Not a style function with space before '('
